Refactor everything
