+++ app/src/main/java/com/AERYZ/treasurefind/main/entry_point/LoginActivity.kt
@@ -23,7 +23,7 @@ class LoginActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_login)
 
         // Choose authentication providers
         val providers = arrayListOf(
