--- app/src/main/java/com/AERYZ/treasurefind/main/ui/Treasures/TreasureViewModel.kt
@@ -1,10 +1,0 @@
-package com.AERYZ.treasurefind.main.ui.Treasures
-
-import androidx.lifecycle.ViewModel
-import com.AERYZ.treasurefind.db.Treasure
-
-class TreasureViewModel : ViewModel() {
-
-    var treasureList = ArrayList<Treasure>()
-
-}
