--- app/src/main/java/com/AERYZ/treasurefind/main/ui/home/HomeViewModel.kt
@@ -1,9 +1,0 @@
-package com.AERYZ.treasurefind.main.ui.home
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-}
