+++ app/src/main/java/com/AERYZ/treasurefind/main/ui/test/TestFragment.kt
@@ -1,25 +1,20 @@
-package com.AERYZ.treasurefind.main.ui.Treasures
+package com.AERYZ.treasurefind.main.ui.test
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
-import android.widget.ListView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
 import com.AERYZ.treasurefind.R
-import com.AERYZ.treasurefind.databinding.FragmentTreasureBinding
-import com.AERYZ.treasurefind.db.Treasure
+import com.AERYZ.treasurefind.databinding.FragmentTestBinding
 import com.AERYZ.treasurefind.main.ui.map.MapsActivity
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
+import com.AERYZ.treasurefind.main.ui.treasureadding.TreasureAddingActivity
 
-class TreasureFragment : Fragment() {
+class TestFragment : Fragment() {
 
-    private var _binding: FragmentTreasureBinding? = null
+    private var _binding: FragmentTestBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -31,17 +26,24 @@ class TreasureFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val treasureViewModel =
-            ViewModelProvider(this).get(TreasureViewModel::class.java)
 
-        _binding = FragmentTreasureBinding.inflate(inflater, container, false)
+        _binding = FragmentTestBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
+        //Map Activity button
         val btn_mapsActivity: Button = root.findViewById(R.id.btn_map)
         btn_mapsActivity.setOnClickListener() {
             val intent = Intent(requireContext(), MapsActivity::class.java)
             startActivity(intent)
         }
+
+        //Map Activity button
+        val btn_treasureAdding: Button = root.findViewById(R.id.btn_treasureadding)
+        btn_treasureAdding.setOnClickListener() {
+            val intent = Intent(requireContext(), TreasureAddingActivity::class.java)
+            startActivity(intent)
+        }
+
         return root
     }
