+++ app/src/main/java/com/AERYZ/treasurefind/main/ui/treasureadding/TreasureAddingActivity.kt
@@ -1,0 +1,12 @@
+package com.AERYZ.treasurefind.main.ui.treasureadding
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import com.AERYZ.treasurefind.R
+
+class TreasureAddingActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_treasure_adding)
+    }
+}
