+++ app/src/main/java/com/AERYZ/treasurefind/main/ui/treasureadding/TreasureAddingViewModel.kt
@@ -1,0 +1,7 @@
+package com.AERYZ.treasurefind.main.ui.treasureadding
+
+import androidx.lifecycle.ViewModel
+
+class TreasureAddingViewModel : ViewModel() {
+
+}
