+++ app/src/main/res/layout/fragment_test.xml
@@ -4,12 +4,20 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.ui.Treasures.TreasureFragment">
+    tools:context=".main.ui.test.TestFragment">
 
     <Button
         android:id="@+id/btn_map"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Map Activity"/>
+        android:text="Map Activity"
+        android:layout_margin="5dp"/>
+
+    <Button
+        android:id="@+id/btn_treasureadding"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Treasure Adding"
+        android:layout_margin="5dp"/>
 
 </LinearLayout>
