+++ app/src/main/res/menu/activity_home_drawer.xml
@@ -5,14 +5,6 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_home"
-            android:icon="@drawable/ic_menu_camera"
-            android:title="@string/menu_home" />
-        <item
-            android:id="@+id/nav_treasure"
-            android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_treasure" />
-        <item
             android:id="@+id/nav_feed"
             android:icon="@drawable/ic_menu_slideshow"
             android:title="@string/menu_feed" />
@@ -20,5 +12,9 @@
             android:id="@+id/nav_profile"
             android:icon="@drawable/ic_menu_slideshow"
             android:title="@string/menu_profile" />
+        <item
+            android:id="@+id/nav_treasure"
+            android:icon="@drawable/ic_menu_gallery"
+            android:title="@string/menu_test" />
     </group>
 </menu>
