+++ app/src/main/res/values/strings.xml
@@ -10,10 +10,9 @@
     <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
 
-    <string name="menu_home">Home</string>
-    <string name="menu_treasure">Treasure</string>
     <string name="menu_feed">Feed</string>
     <string name="menu_profile">Profile</string>
+    <string name="menu_test">Test</string>
 
     <string name="onboarding_title1">Ahoy, Matey!</string>
     <string name="onboarding_title2">Shiver me timbers!</string>
